0.6.0 - Unified Mem, DLA Core, JIT, TRT Preproc, Demos (#51)
* Add TRT Preprocessor, engine cache, better inspect output, more verbose flags, add debug flag to TRTEngine (#41)

* script for making preproc simple onnx

* Add debug flag for intermediate stream sync, more verbose flags, beginnings of trt preprocessing engine

* old lock in CUDAPreproc removed

* Add dumb AI gen hook for now, no integration yet many CI errors

* further work on trt preproc, create caching system, build_engine can use cache. hook passes linter

* engine compilation, need to add new execution mode for TRTEngine

* working at runtime, seg fault during tests, incorrect results

* dont cache the engine for trt preprocessing during testing, formatting, fixed tests

* extra test for cuda v trt compare

* add trt preprocessor tests

* add trt preproc tests to other yolo versions

* refactor preprocessors to split into many files

* cuda results overall tests pass now, error in re-allocation which re-allocated blocks

* replicate fix in trt preprocessor

* remove final stream sync on cleanup to prevent segfault, need to handle better later. Better test analysis for trt preproc

* trt preprocessing working, now need to optimize

* add call method back to trtpreprocessor, still failing generating bounding boxes

* fully working and tested TensorRT micro-engine based preprocessor

* pass mypy CI

---------


* Lazy Load libnvinfer plugins and YOLO preprocessors (#42)

* defer reading kernel file until Kernel init

* Lazy load nvlibinfer plugins

* Lazy load the CUDA/TRT preprocessors, add corresponding tests

* resolve some mypy errors

* resolve type error in tests

---------


* Add DLA Core Specification (#43)

* specify dla_core in create_engine and TRTEngine

* tested on jetpack6

* inital impl for supporting dla_core assignment in other TRTEngine abstractions

* extend dla_core specification to CLI

---------


* Benchmark alternative preprocessing methods separate (#44)

* Add trtutils alt-preprocessing methods

* add tensorrt standalone benchmark

* update titanrtx data and plots

* update orinagx64gb data and plots

* update tags in data, update plot to skip frameworks not present

* new benchmark docs from titanrtx and orinagx

---------


* Add JIT (#45)

* initial implementation

* jit installation tag

* modifications to yolo CLI demo

* convert to ms and round at start

* aggregate mean times for each stage

* ci

* use global logger

* Webcam demo (#46)

* scaffold of webcam demo

* initial version, pending tests

* source always int

* Multi stream demo (#47)

* first draft of multi_stream demo

* Add single stream version

* working streaming with GPU + DLA0, adding dla1 causes contention instead of allowing faster processing, must debug

* update multi_stream_demo with better DLA assignment

* Add readmes to demos, todo: fill in

* README, requirements, and CI checks demos using ruff

---------


* Faster SST Kernel (#48)

* have detection info output be from verbose flag yolo cli

* work on shared memory tile optimization for based on transpose kernel article

* Add test code for new fast impl of SST kernel

* swap sst for sst_opt

* WIP, tests resolve and fast sst compiles and runs

* pass ci

* pass CI, fully pass tests

---------


* Use Managed CUDA Memory (#49)

* Always allocate pagelocked, add host/device pointer get func

* update new mem function from logger/lock changes

* pagelocked mem propagation flag working, test for parity/performance

* pagelocked and mapped memory working successfully, speedup higher. Need benchmark to assess slowdown at large input size from np.copyto

* Managed/pagelocked CUDA memory used in TRTEngine/YOLO, perf unit tests, correctness verified

* unit tests for pagelocked YOLO perf

* CI on src/tests, slight changes to engine mock performance

---------


* 3080ti data

* orinagx 64gb data

* add new 3090 data

* partial xaviernx data

* orinagx-32gb

* new orinnx-16gb data

* new orinnano data

* benchmark script run changes

* resolve always overwrite issue

* Issue warning if trying to use TRT preprocessor and fallback to CUDA on Xavier

* titanrtx data

* CI passing, update CI requirements

* orinnx full

* orinagx 32 full

* Separate Unified Mem from Pagelocked (#50)

* added unified_mem flag to all classes, auto-determine from IS_JETSON, perf test fixed regression on RTX

* update titanrtx data, todo 3080, 3080ti

* 3080ti rerun

* add 3090 data

* xaviernx ultralytics data

* full CI pipeline ready, all tests pass, docs updated

* Bump version 0.5.0 -> 0.6.0

* update changelog

* incorrect verions for swapping preproc tests

* resolve memcpy error from non-pagelocked memory, resolve array equality checks in yolo_preproc tests

---------

+++ src/trtutils/jetson/_benchmark.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from jetsontools import Tegrastats, filter_data, get_powerdraw, parse_tegrastats
+from jetsontools import TegraStats, filter_data, get_powerdraw, parse_tegrastats
 
 from trtutils._benchmark import Metric
 from trtutils._engine import ParallelTRTEngines, TRTEngine
@@ -39,8 +39,10 @@ def benchmark_engine(
     iterations: int = 1000,
     warmup_iterations: int = 50,
     tegra_interval: int = 5,
+    dla_core: int | None = None,
     *,
     warmup: bool | None = None,
+    verbose: bool | None = None,
 ) -> JetsonBenchmarkResult:
     """
     Benchmark a TensorRT engine on a Jetson device.
@@ -58,9 +60,15 @@ def benchmark_engine(
         The number of milliseconds between each tegrastats sampling.
         The smaller the number, the more samples per second are generated.
         By default 5 milliseconds between samples.
+    dla_core : int, optional
+        The DLA core to assign DLA layers of the engine to. Default is None.
+        If None, any DLA layers will be assigned to DLA core 0.
     warmup : bool, optional
         Whether to do warmup iterations, by default None
         If None, warmup will be set to True.
+    verbose : bool, optional
+        Whether ot not to output additional information to stdout.
+        Default None/False.
 
     Returns
     -------
@@ -69,7 +77,13 @@ def benchmark_engine(
 
     """
     if isinstance(engine, (Path, str)):
-        engine = TRTEngine(engine, warmup_iterations=warmup_iterations, warmup=warmup)
+        engine = TRTEngine(
+            engine,
+            warmup_iterations=warmup_iterations,
+            warmup=warmup,
+            dla_core=dla_core,
+            verbose=verbose,
+        )
     else:
         if warmup:
             for _ in range(warmup_iterations):
@@ -80,16 +94,16 @@ def benchmark_engine(
     raw: dict[str, list[float]] = {metric: [] for metric in metric_names}
 
     # pre-generate the false data
-    false_data = engine.get_random_input()
+    false_data = engine.get_random_input(verbose=verbose)
 
     # create temp file location for data to go
     temp_file = Path(Path.cwd()) / "temptegra.txt"
     # store the start/stop times of the engine execution
     start_stop_times: list[tuple[float, float]] = []
-    with Tegrastats(temp_file, interval=tegra_interval):
+    with TegraStats(temp_file, interval=tegra_interval):
         for _ in range(iterations):
             t0 = time.time()
-            engine.mock_execute(false_data)
+            engine.mock_execute(false_data, verbose=verbose)
             t1 = time.time()
             raw["latency"].append(t1 - t0)
             start_stop_times.append((t0, t1))
@@ -98,7 +112,8 @@ def benchmark_engine(
     tegradata = parse_tegrastats(temp_file)
 
     # delete the temp file
-    temp_file.unlink()
+    if temp_file.exists():
+        temp_file.unlink()
 
     # filter the data by actual times during execution
     filtered_data, per_inference = filter_data(tegradata, start_stop_times)
@@ -133,20 +148,21 @@ def benchmark_engine(
 
 
 def benchmark_engines(
-    engine_paths: Sequence[Path | str],
+    engines: Sequence[TRTEngine | Path | str | tuple[TRTEngine | Path | str, int]],
     iterations: int = 1000,
     warmup_iterations: int = 50,
     tegra_interval: int = 5,
     *,
     warmup: bool | None = None,
     parallel: bool | None = None,
+    verbose: bool | None = None,
 ) -> list[JetsonBenchmarkResult]:
     """
     Benchmark a TensorRT engine.
 
     Parameters
     ----------
-    engine_paths : Sequence[Path | str]
+    engines : Sequence[TRTEngine | Path | str | tuple[TRTEngine | Path | str, int]]
         The engines to benchmark as paths to the engine files.
     iterations : int, optional
         The number of iterations to run the benchmark for, by default 1000.
@@ -164,6 +180,9 @@ def benchmark_engines(
         Useful for assessing concurrent execution performance.
         Will execute the engines in lockstep.
         If None, will benchmark each engine individually.
+    verbose : bool, optional
+        Whether ot not to output additional information to stdout.
+        Default None/False.
 
     Returns
     -------
@@ -172,6 +191,21 @@ def benchmark_engines(
         If parallel was True, will only contain one item.
 
     """
+    temp_engines: list[Path | TRTEngine] = []
+    dla_assignments: list[int | None] = []
+    for engine_info in engines:
+        engine: TRTEngine | Path | str
+        dla_core: int | None = None
+        if isinstance(engine_info, tuple):
+            engine = engine_info[0]
+            dla_core = engine_info[1]
+        else:
+            engine = engine_info
+        if isinstance(engine, str):
+            engine = Path(engine)
+        temp_engines.append(engine)
+        dla_assignments.append(dla_core)
+
     if not parallel:
         return [
             benchmark_engine(
@@ -179,15 +213,19 @@ def benchmark_engines(
                 iterations,
                 warmup_iterations,
                 tegra_interval,
+                dla_core=dla_core,
                 warmup=warmup,
+                verbose=verbose,
             )
-            for engine in engine_paths
+            for engine, dla_core in zip(temp_engines, dla_assignments)
         ]
 
     # otherwise we need a parallel setup
-    engines_paths = [Path(ep) for ep in engine_paths]
-    engines = ParallelTRTEngines(
-        engines_paths,
+    trt_engines = ParallelTRTEngines(
+        [
+            (ep, dc) if dc is not None else ep
+            for ep, dc in zip(temp_engines, dla_assignments)
+        ],
         warmup_iterations=warmup_iterations,
         warmup=warmup,
     )
@@ -197,17 +235,17 @@ def benchmark_engines(
     raw: dict[str, list[float]] = {metric: [] for metric in metric_names}
 
     # pre-generate the false data
-    false_data = engines.get_random_input()
+    false_data = trt_engines.get_random_input()
 
     # create temp file location for data to go
     temp_file = Path(Path.cwd()) / "temptegra.txt"
     # store the start/stop times of the engine execution
     start_stop_times: list[tuple[float, float]] = []
-    with Tegrastats(temp_file, interval=tegra_interval):
+    with TegraStats(temp_file, interval=tegra_interval):
         for _ in range(iterations):
             t0 = time.time()
-            engines.submit(false_data)
-            engines.retrieve()
+            trt_engines.submit(false_data)
+            trt_engines.retrieve()
             t1 = time.time()
             raw["latency"].append(t1 - t0)
             start_stop_times.append((t0, t1))
@@ -216,7 +254,8 @@ def benchmark_engines(
     tegradata = parse_tegrastats(temp_file)
 
     # delete the temp file
-    temp_file.unlink()
+    if temp_file.exists():
+        temp_file.unlink()
 
     # filter the data by actual times during execution
     filtered_data, per_inference = filter_data(tegradata, start_stop_times)
